refactor: organize tracking package
