+++ glassflow-api/internal/service/usage_stats_collector.go
@@ -189,7 +189,7 @@ func (m *UsageStatsCollector) sendPipelineMetrics(ctx context.Context, pipeline 
 	}
 
 	m.usageStatsClient.SendEvent(ctx, "pipeline_metrics", "api", map[string]interface{}{
-		"pipeline_id_hash": usagestats.HashPipelineID(pipelineID),
+		"pipeline_id_hash": usagestats.MaskPipelineID(pipelineID),
 		"ingestor_metrics": ingestorMetrics,
 		"dedup_metrics":    dedupMetrics,
 		"join_metrics":     joinMetrics,
